Properly sort packaging and move code into a src dir
https://hynek.me/articles/testing-packaging/ thanks Hynek!
