+++ src/python_template/skeleton.py
